Updated DeviseGenerator to respect singular tables.
If ActiveRecord::Base.pluralize_table_names is set to false,
the table for AdminUser will be admin_user, but the route
generated by Devise will still be admin_users.  However, the
Active Admin route generation assumed the route name would
always be the same as the table name when doing a gsub in the
routes file.  Since the gsub didn't match the trailing 's' in
the route name, you ended up with the following in routes.rb:

  devise_for :admin_user, ActiveAdmin::Devise.configs

This resulted in a NoMethodError because of the leftover 's'
on the end of the line:

  undefined method `configs' for ActiveAdmin::Devise:Module (NoMethodError)

Switching from table_name to plural_table_name in the generator
ensures it matches the complete route name regardless of the
value of ActiveRecord::Base.pluralize_table_names.