Cucumber tests for default_action changes
I have not written any unit specs as there is no index_as_table_spec. I won't be
able to go through the effort to figure out how to create the setup for such
specs, but I would be happy to write a spec if and when there is an
index_as_table_spec up and running.