No need to delete the :if from the options hash.
We need to be very aware of when we use Hash#delete. It usually doesn't
cause side effects in development because we're reloading resources
over and over, but in production it can cause us to delete data the
on the first usage. Usually we just want a reference to the value
instead of actually removing it from the hash.