Refactor Naming for Config and Resource!
Note: The documentation says that when registering a resource you can
set its name doing:

```
ActiveAdmin.register Post, :as => "Article"
```

The code however was made so that the :as resource would get a
plural. For instance:

```
ActiveAdmin.register Post, :as => "Articles"
```

I fixed this. It might be worth mentioning in the release notes.