Update assets_generator to place assets in vendor assets directory.
This is the correct place for assets that are added to an app. Per the Rails documentation "vendor/assets is for assets that are owned by outside entities, such as code for JavaScript plugins." This also fixes a problem where using the sprockets directive require_tree in application.css automatically includes the active admin CSS everywhere - organizing assets into vendor like this will load them only on active admin pages.