ActiveAdmin::Resource looks up classes at runtime using their name
To deal with the reloading issues in #870, we now store reference to the
resource class as a string and constantize it each time we need it.

Also added a new cucumber profile called "class-reloading" which does
not cache classes. Since Active Admin should always work as expected in
development with Rails reloading, we should continue to grow scenarios
that include the '@requires-reloading' tag.