Add any object that repsonds to to_s to the DOM in Arbre blocks.
Before this commit:

  index do
    column("My Column"){ 1 }
  end

Would result in an empty column. Now the object is appended to the DOM
and will result in a colum of "1"s.

The return value is ignored if the user has already appended other
elements to the DOM. This might trip up some users, however it is more
likely that you are either going to build up a DOM structure or return a
simple value, not both.

Thanks for the pull requests from @bobbytables and @utkarshkukreti

References isses #470, 154