fix a bug in AA::Inputs::FilterSelectInput#method
This method takes something like "author_id" and returns "author". It
was failing for the edge case "external_identity_id" because of the
extra "_id" in the middle of the string - i.e. it was returning
"externalentity".

Change the way the substitution is done so that it only matches at the
end of the string.