Completely refactored the menu system to work with i18n
Menu's are now generated on demand using the MenuBuilder instead of
being created as the resources are registered. MenuItem can now store
labels as procs, which allows us to properly i18n all the menu labels
at runtime.

ActiveAdmin::Resource's are now in charge of building their own MenuItem
objects.

The controller now stores the instance of the currently selected tab in
@current_tag instead of just a string label.