Add formating ability to the Dashboard
A table is used to display the sections on the dashboard.
This table has 3 columns (hard-coded).

This is an over-ride on the render_sections method
in the ActiveAdmin::Views::Pages::Dashboard class.

The new method does two things:

  - add the 'colspan' attribute if it was definde as
    an option of the section
  - format the table to be 12-column wide

By default, the colspan is set to 4 to keep the
previous layout of three columns.