singularize "New ______" Button
Hey, I'm not sure if this is the right spot for it... but I'm getting a         
'New Deadlines' button in my admin panel instead of a 'New Deadline'            
one. Adding this change with `bundle open activeadmin` fixes the problem        
without any obvious side-effects.  Note: I'm not sure if anywhere else          
needs to be changed. 