move from Metasearch to Ransack
Note that while Ransack renames many query methods ("predicates"), this
commit creates aliases for the more plain-english predicates so they
still work as expected.

To see the full list of Ransack predicates, check out
`lib/ransack/constants.rb` in the Ransack source code.

This commit also removes polymorphic foreign keys from the default
list of filters for a resource, since Ransack seems to barf on them. We
don't have a good filter UI for them anyway, so it's not a huge issue.