make `pretty_format` compatible with Bignum
In particular, Bignum doesn't support `instance_exec` because it
doesn't have a singleton class. Previously `pretty_format` would pass
the number to `display_name`, which would default in running it against
the DISPLAY_NAME_FALLBACK proc.

Note that this should also be fixed in `display_name` itself, but any
attempt to do so hanged, consuming all memory on my computer :<