normelize action in form view title
Now the action is mapped from `create` to `new` and `update` to `edit`. It doesn't make sence to show a special title for the `update` or `create` action, they should have the same title as `edit` or `new`. The title of `update` and `create` is only shown if a validation fails.