Streamline table styles, use better defaults
Tables are the most important part of ActiveAdmin as it’s the
components that is most widely used to display data. Keep presentation
clean and consistent, same padding, colors, etc. We can fine tune but
this brings it down just to the necessary styles without being verbose.
Further improvements: removing even/odd classes as this can now be done
using solely CSS.