Update blank_slate component and styles
Now works pretty much the same as a Panel component. This way we can
use `blank_slate` with a block and specify whatever elements we like
rather than passing a content string (although that is still
supported). Now our dashboard template is much cleaner as we use the
built in component rather than hard code CSS and styles.

Apply button style to blank slate action on index.

Display text in a paragraph so button is below it but spaced out
using our primary button style.

Further configure styles in variables.