Add Contributor Covenant as our code of conduct
The purpose of adopting this code of conduct is to signal to potential contributors that this is a project where they won't be abused or attacked in any way. It also signals to everyone else that such behavior will not be tolerated.

This COC is copied from http://contributor-covenant.org/

I'll leave this PR open for a week, in the case that people have concerns with the wording.