allow CSV pagination to be customizable
This fixes the performance issues that #3375 introduced by switching
from `find_each` to Kaminari, since Kaminari has to do a SQL count. For
those that want that behavior they can enable it by setting
`paginate_with: :kaminari`, but the fastest option should be the
default for Active Admin.

I chose this approach instead of #4082 because databases don’t
guarantee any particular order when requesting records, so it’s
possible that the resulting CSV won’t include all records.

In this commit I removed the `batch_size` method. Ideally it should be
a config option, but I didn’t have the time / motivation to write tests
for it. Currently both pagination methods use a 1000 record page size.

This commit refactors the tests that were creating their own
controllers, to reduce duplication.

While writing this I came across a bug with the byte order mark / BOM
character. Our Cucumber tests had set it to an empty string, which
would cause the unit tests to fail. Turns out we weren’t encoding that
character. I updated the code to fix that, but I didn’t think it was
important enough to write a separate test for.