Sort load path files in ruby
Different OS/OS Versions return a different ordering for the same
directory contents.  Because of this, we can't rely on Rail's
autoloading to load the files in the same order we expect in different
environments, which can lead to bugs between development and production.

By first sorting the result of Dir["#{path}/**/*.rb"], we make sure that
the files are returned in the same order for each environment,
regardless of OS inconsistancies.

We are also uniq'ing the files returned after, which retains the order
defined in the `load_paths`, but removes any duplicates from entire
list.  It is possible that doing that will lead to potentially
unintendened concequences for the user, so that might be something that
should be removed in the future.