Not override but add behaviour to existing member or collection action if the same name specified
If somebody wants to create an action, that handles multiple HTTP verbs, they might write code like this:

          ActiveAdmin.register Post do
            member_action :action_name, method: :get do
              # do something on GET request
            end

            member_action :action_name, method: :post do
              # do something on POST request
            end
          end

But this will not work as expected. There is no magic  )
The second declaration will override the first one.

Furthermore, there will be 2 routes generated (GET, POST), that is strange.

If there are two routes, why one behaviour?

At least, If action is overrided then previous route should be removed.

This commit adds a lot of magic and automatically turns the above code into something like this:

          ActiveAdmin.register Post do
            member_action :action_name, method: [:get, :post] do
              if request.post?
                # do something on POST request
              else
                # do something on GET request
              end
            end
          end

In case when action name and http verb both the same as previous declared,  the latest declaration overrides the privious ones.

This works for both member and collection actions and with any combination of actions and http verbs.