Add i18n content for date range filter
I think this was broken in 1.0.0.pre3 because of the change in issue
4505. The Ransack configuration was changed to add gteq_date and
lteq_date as predicates but ActiveAdmin's i18n keys were not updated
with this content. Ransack does not provide default values. Without the
i18n content you get translation errors in the sidebar search status on
a results page.