Test against 2.4.0
Only one spec to be fixed, and it's actually a bug fix brought to you by
the awesome ruby 2.4.0 unicode case mappings! ;)

The new Travis matrix is as follows:

* Ruby 2.4.0 + Rails {3.2,4.0,4.1} combination segfaults. Since those
Rails versions are already EOLed, I think it's pointless to try to
bother with them, since the segfaults probably comes from inside Rails
and they will no longer be patched.

* Ruby 2.4.0 + Rails 4.2 fails, but Ruby 2.4 support will be added in
the next release. So I add it to `allowed_failures` but we should be
able to move it back to the regular matrix once 4.2.8 is released.