Allow passing an Array to collection_size
This allows you to use not only a ActiveRecord::Relation but also an Array as collection for code that calls ActiveAdmin::Helpers::Collection.collection_size.

Passing an Array might be needed in case of creating a table on a custom page with a paginated array as collection.
See https://github.com/activeadmin/activeadmin/issues/4423