Add indeterminate support if some checkboxes are checked
A nice visual enhancement that shows the toggle all checkbox in an
indeterminate state which is visual only (similar to GitHub for Mac or
Gmail web). I have a jQuery plugin I’m working for this behavior so I’m
already familiar with what to add. For more info:
https://css-tricks.com/indeterminate-checkboxes/

Performance improvements as there is no need to loop over checkboxes to
set checked property, the prop method will work on a collection.
Further the loop was needed to trigger didChangeCheckbox method for the
table-checkbox-toggler sub class but the better way is for that
subclass to override the didChangeToggleAllCheckbox otherwise you’re
calling didChangeCheckbox for every checkbox and running DOM queries
and updates that are unnecessary. Existing behavior preserved as all
tests are passing.