Add a sidebar with attributes_table_for sample
To fit something there that made sense, I thought the author data in
the sidebar was best so I moved that there and in its previous place I
just have it show category which is a has one relationship so also good
for an attributes_table_for sample in the panel. Trimmed down the tags
table and renamed some columns.