Create a custom page and call it Kitchen Sink
This is useful since it shows how to create a custom page that isn’t
tied to a model but also as its name suggests, a dumping ground for
various misc things we want to test out that don’t really fit easily
elsewhere. Things like tabs and various basic HTML that we want to see
it rendered so we can see how AA styles them.