Update app config/requires to match Rails bug template scripts
We should stay consistent with how Rails sets up their action
controller/view gem script template
https://github.com/rails/rails/blob/master/guides/bug_report_templates/a
ction_controller_gem.rb as we need that to properly test ActiveAdmin.

Use preferred ActionDispatch::IntegrationTest for tests. This is best
since it goes through the full Rails stack and we need that coverage
for ActiveAdmin since the view is a big portion of it.

Instead of YourModels, just create User model. Its weird having
examples like that. I think its clearer having a real world example
(Which the Rails templates do the same) so its easier to get
started and see what’s going on as people are familiar with that.