Streamline content so its readable in preview
Painful to use comments here. Best to remove it like in other repo
issue templates so it renders nicely in Preview with links enabled.
Focus on directing users to the contributing guidelines so that way
they provide us with an executable script. These are incredibly useful
and can be coded faster than a sample app. Wish I knew we had this
setup before as I’ve submitted Rails issues with their script template
and its great.