Add default templates for Kaminari
These are what we style and fix earlier issues where the host app has a
custom kaminari theme but the admin doesn’t. Since the admin uses the
default, we pick up what the host app sets. Ideally, we should have
this in house so we can style as we like which will be updated again in
#3862 anyway. This is just how Kaminari is set up and I don’t think its
wrong. We should just handle this like we do Devise. We should own it
so we can style as we like. We won’t want to get rid of Kaminari, its
great at what it does!