Update per page to es6
Tested all functionality and works great even with existing or no query
params. I adjusted the regex to account for the preceding “?” character
in the query string rather than just doing `substring(1)` at the end of
`window.location.search` which isn’t clear at first why its there. I
preferred to keep everything in the regex as it makes sense since the
character is part of the string.