Fix comments on decorated object in Rails v5.2
When upgrading from Rails 5.0 to 5.2, the call to #active_admin_comments started failing for the view action of one of my models.

This model has a decorator set using `decorate_with`, using a PORO decorator as described in the second paragraph in the decorator documentation: https://activeadmin.info/11-decorators.html. I have not checked if this problem exists when decorating with draper.

I tracked the problem to this line, where the resource itself was being passed as a parameter, leading to unexpected behaviour when an attempt to cast the decorated object in ActiveRecord::ConnectionAdapters::Quoting.type_cast raised a TypeError.

I suspect the problem could also be fixed by adding some missing methods to the PORO decorator, but it seems like a better fix to simply pass the actual ID at this line in stead of the entire resource, eliminating the risk of similar bugs in the future.