Remove unnecessary requires
Move all the common stuff to an `.rspec` file. The explicit requires are
unlikely to be a problem but they can be annoying in some situations.
For example, if you introduce a syntax error in the helper, you will get
it printed hunders of times, one per file, since the require never
succeeds and it's retried everytime a new spec file is loaded.