Move lint tools to a "lint" group in Gemfile
The only difference is that custom groups are not required by default.
Since this tools are explicitly required when used, we can do this. The
main reason why I do this is that on jruby lots of warnings from rubocop
are spit during test app generation, like the following:

```
warning: parser/current is loading parser/ruby25, which recognizes
warning: 2.5.1-compliant syntax, but you are running 2.5.0.
warning: please see https://github.com/whitequark/parser#compatibility-with-ruby-mri.
```

We skip the with this trick.