Make jruby hangs appear earlier
The `parallel_tests` gem has this feature where it will output a dot
every 60 seconds to avoid CIs that fail when no new output has not been
displayed for a while.

Our jruby build is hanging randomly and when that happens, this feature
makes it take hours for the job to fail.

In this commit, I'm essentially disabling this feature, by giving the
heartbeat a high value, so that we actually get these failures earlier.