Workaround turbolinks flaky failure
Apparently, when loading the posts index page, the "View" link is
available without turbolinks behaviour attached for some time. I can
even reproduce this on the test application by navigating to an index
page, reloading the page, and quickly clicking on a "View" link: nothing
happens.

We would need to investigate this issue more deeply, but for now I'm
changing the feature to navigate directly to the page where the tabs
are, so that it doesn't hit this issue.