Improved polymorphic routes generation.
If we are going to use the configured `belongs_to` resource to generate the path part, is better to compare the given record with that class, instead of doing it with the `parent` class.
This change shouldn't affect when working with just one `belongs_to` sentence, but it should fix the paths generation with multiple `belongs_to`.

Added tests for polymorphic routes with multiple belongs_to

Changelog entry added