Disallow extra spacing
Some people like this code because it's prettier. I don't, because it
makes diffs harder to read. Sometimes it forces you to change a bunch of
lines just to keep the "prettyness" when you just want to really fix a
single line. In my opinion, this hides meaningful changes inside
stilistic ones.

For example, say I want to fix a bug caused by a typo in one of the keys
in a big hash.  I used `deployments` but the actual key ìs `deployment`.

```ruby
{
  :clean       => true,
  :deployments => true,
  :frozen      => true,
  :"no-cache"  => true,
  :"no-prune"  => true,
  :path        => "vendor/bundle",
  :shebang     => "ruby27",
  :system      => true,
  :without     => "development",
  :with        => "development",
}
```

If using the "pretty alignment" style, I need to edit all lines in the
hash just to make the fix, in order to keep the style, so the diff would
look like:

```diff
{
- :clean       => true,
- :deployments => true,
- :frozen      => true,
- :"no-cache"  => true,
- :"no-prune"  => true,
- :path        => "vendor/bundle",
- :shebang     => "ruby27",
- :system      => true,
- :without     => "development",
- :with        => "development",
+ :clean      => true,
+ :deployment => true,
+ :frozen     => true,
+ :"no-cache" => true,
+ :"no-prune" => true,
+ :path       => "vendor/bundle",
+ :shebang    => "ruby27",
+ :system     => true,
+ :without    => "development",
+ :with       => "development",
}
```

In my opinion this is both incovenient, and also hides the fix. With a
no spacing style, it's simpler:

```diff
{
  :clean => true,
- :deployments => true,
+ :deployment => true,
  :frozen  => true,
  :"no-cache" => true,
  :"no-prune" => true,
  :path => "vendor/bundle",
  :shebang => "ruby27",
  :system => true,
  :without => "development",
  :with => "development",
}
```