Update mdl to then update kramdown for security fix (#6391)
Dependabot couldn't create an update for kramdown because the mdl gem depends on it but an older version. We have to update both to get the security fix. I'm not sure why we haven't received Dependabot updates for the mdl gem when it has several new releases. We do get Dependabot updates for the other gems in the lint group so something else may be wrong.

https://github.com/activeadmin/activeadmin/network/alert/Gemfile.lock/kramdown/open

https://github.com/advisories/GHSA-mqm2-cgpr-p4m6

When first working on this update, we ran into an error on mdl v0.10.0 as reported here: https://github.com/markdownlint/markdownlint/issues/339 but was address in v0.11.0 which this now uses. The lint:mdl task is now running successfully on v0.11.0.