Automatically create a release branch from release scripts
Current scripts tell you to start off the default branch, and release
scripts don't do any branch switching, so they would commit to the
default branch and tell you to push after that. We don't want that, but
push to a non default branch and create a PR from it.