Strip "invisible" trailing spaces from PR title (#6476)
If a PR title include trailing spaces, the generated CHANGELOG entry has
an extra space before the final dot:

```
* Fix input string filter no rendering dropdown input when its column name ends with a ransack predicate . [#6422] by [@Fivell]
```

This kind of formatting little issues can be easily corrected by
changing the PR title, but in this particular case Github doesn't
display this trailing space, so it's hardly impossible to notice. We
can fix it when generating the changelog entry instead.