Include print styles in main stylesheet (#6922)
No need to do this separately which makes an extra http request, plus
this has the benefit of delaying print styles until needed. I’ve been
doing this for a long time now and as its been well supported to
include print styles using a media block. The HTML5 Boilerplate has
stressed this as a good default approach for many years now.
https://github.com/h5bp/html5-boilerplate/blob/c9f88cc79d380150e4e2e39e3c2deccfe8aa3d8c/dist/css/style.css#L189-L193

Co-authored-by: Javier Julio <jjfutbol@gmail.com>