Add test for helper method with kwargs call in form
When a helper method accepts kwargs and they are given, this causes a
failure in Ruby 3.0 with arbre 1.4.0 or lower.

> wrong number of arguments (given 2, expected 1) (ActionView::Template::Error)
> ./tmp/test_apps/rails_61/app/helpers/time_helper.rb:3:in `format_time'

If the keyword arguments are all optional and NOT given, then no error occurs.

Fixed in arbre itself but we didn't manage to write a test directly in
there, so adding a higher level one here.