Migrate code coverage to Codecov (#7853)
* Migrate i18n_spec.lint.rb to regular spec

* Migrate local_spec.lint.rb to regular spec

* Migrate changelog_spec.lint.rb to regular spec

* Migrate gemfiles_spec.lint.rb to regular spec

* Run against current gemfile set or env

No need to run this against all gemfiles, although we can continue to do that, since we set the BUNDLE_GEMFILE in the CI matrix or when running tests locally, we can just test against that.

* Migrate gemspec_spec.lint.rb to regular spec

We'll try running this as regular spec (parallel tests) to see if its fine, if not we'll flag it with a changes_filesystem filter.

* Migrate js_bundle_spec.lint.rb to regular spec

* Update js_bundle test to yarn install

Since this was run as part of the lint rake task, it was already running yarn install as a prerequisite due to gherkin lint. So we'll need it here. For consistency use binstub bin/yarn instead.

* Remove lint simplecov command

We are consolidating these specs into the regular test suite. Managing these separately in CI is too involved overall. No need to treat those lint specs as special, they are just specs.

* Remove lint rspec task

These are now migrated to normal specs (parallel tests). Managing these separately in CI is too involved overall. No need to treat those lint specs as special, they are just specs.

* Use simplecov-cobertura for Codecov

* Update CI to use Codecov upload

* Delete Code Climate integration

Delete prepare_coverage binstub since this was just needed for Code Climate

* Remove kramdown from changelog spec

We'd have to reconfigure the gemfiles so kramdown is part of test group. For this single test case though, I don't think it's worth the change but we can always revisit and add back in. The main focus is to migrate the lint specs to just be normal specs.

* Add i18n-tasks and i18n-spec to all gemfiles

* Update coverage badge