Optimize CI (#7897)
* Update daily workflow to use root gemfile only

This was failing since we don't have a rails_70 gemfile directory. That's assumed as the default in the root gemfile. We'll just use that for now.

Update workflow id to jruby_tests.

Only run workflow on the activeadmin repo, do no run on forks.

* Migrate jruby lint job to daily workflow

All this did was run the bug report template but since we run the full tests on jruby in the daily workflow just makes sense to run it there and not part of regular CI.

* Migrate gherkin-lint to optimized workflow

No need to run gherkin-lint from Ruby. We can run it in an optimized workflow that will only run when applicable.

* Update simplecov to filter spec directory

* Bundle update all outdated dependencies

For some reason Dependabot is downgrading some dependencies in the rails_61_* gemfiles but I don't know why.

* Only set COVERAGE on test run steps

Since we build the test app, the COVERAGE variable exists then and is triggering simplecov when we don't want it too.

* Add github-actions to Dependabot config

* Adjust codecov project threshold

* Remove git setup step in lint workflow

In the CI workflow, indicate that get setup is needed for the test app setup step.