Additional CI and Dependabot improvements (#7900)
* Run daily workflow at noon

* Migrate bug report template to optimized workflow

This extracts the bug report template check from the lint workflow to its own workflow so it only runs when being updated and now runs on a daily schedule to ensure it completes successfully since we may not be changing it much at all.

* For gemfiles/* use monthly interval

These shouldn't change as much (mostly with rails) but we aren't as interested in updating so frequently. Also issues with Dependabot downgrading dependencies that it shouldn't be so we want to review without auto merging for now. Just too hard to manage for ActiveAdmin due to the number of dependencies we declare.

* For npm use monthly interval

Despite this being the root package and we want to keep up to date, these packages tend to be noisy since they have such frequent releases. Let's try updating on a monthly basis.

* Delete mergify.yml

Dependabot is downgrading dependencies it shouldn't be but its hard to manage with several gemfiles and all being auto merged. If it weren't for the wrong downgrading issue, I don't mind auto merging but its hard to manage so let's try it without for now. We'll leave it on for Arbre/Inherited Resources since they have far fewer dependencies compared to ActiveAdmin.

* Don't use bundler to run bug report template

* Add concurrency to tests CI