Minor CI Updates (#7902)
* Simplify ci concurrency

We can just use the actual event name to clarify intent.

* Add concurrency and cancel on pull request only

* Disable daily workflow due to failures

I believe this jruby test started failing after upgrading nokogiri/racc due to the error. It's very time consuming to keep this up to date and passing with minor updates. For now we'll disable and see if others can resolve, if not we'll just remove this workflow.

* Update comment for schedule cron value

To clarify that these times are in UTC per docs.

* Reset codecov project default threshold

We should have resolved most Codecov issues regarding files that should be ignored so this should be ok to maintain going forward.