Remove webpacker/turbolinks builds
This only removes the CI builds for webpacker and turbolinks which are both long archived. The webpacker build we had to stop because it started to fail and we couldn't figure out a solution.

The functionality for each will remain for now while we transition to v3 and a new setup using cssbundling-rails and jsbundling-rails.