Remove metasearch backport, use ransack defaults
This closes #3511 and is updated branch with no merge conflicts. Instead of fixing each conflict, since the original PR is very old, started from scratch making each change and running the tests.

This removes backports for a gem that was deprecated in favor of ransack. We want to reduce our footprint and just rely on ransack's defaults and docs.