Disable simplecov-cobertura require
Since this didn't have an explicit `require: false` this still triggered Simplecov locally when running tests. We can still get code coverage locally to generate, we just want it to be opt-in as there is a time cost to it. With this fix we just need to prepend any command with `COVERAGE=true` and code coverage will be reported.