Use display name fallback if blank display name result (#6342)
* Use display name fallback if method value is blank

This addresses an issue where if a model has a method that is one in the list of configured display_name_methods but it returns a blank value, then the display name will always come back blank. The display name fallback was only used if a model did not have one of the configured display names but it would be helpful to always rely on the display name fallback even if there is no return value otherwise this will cause side effects like this test case in its original form shown, the link had no value and so it wouldn't even appear to be clickable in a browser.

* Use display name fallback if method value is blank

This addresses an issue where if a model has a method that is one in the list of configured display_name_methods but it returns a blank value, then the display name will always come back blank. The display name fallback was only used if a model did not have one of the configured display names but it would be helpful to always rely on the display name fallback even if there is no return value otherwise this will cause side effects like this test case in its original form shown, the link had no value and so it wouldn't even appear to be clickable in a browser.