Remove link rake task
All except for one was to do with whitespace. We want to reduce code footprint and use off-the-shelf tools for basic checks like these. Since they aren't critical we are fine without replacements other than what eslint and rubocop already offer.

The sass-rails lint was for webpacker but we have already removed webpacker support. We won't be using any images or asset URLs in the new v4 approach for assets.