Prep Tailwind CSS migration part 2 (#8036)
* Remove old base, shadows, gradients, print styles

* Replace <body> dynamic classes with data attributes

* Update Panel and Sidebar Section

* Update page base classes and ids

* Remove old status_tag styles

These have been replaced.

* Replace datepicker with date input type

* Update Pagination

This also updates the locale keys for pagination to be less wordy and long. Will follow up to extract the Tailwind CSS classes.

* Footer

* Update TitleBar

The action items are rendered directly into the container so no need for multiple span wrappers with a component which are undesired anyway. The DSL is meant to easily add what action item you want vs implying it has to be a link and styled a certain way.

* Forms and Filters

* DropdownMenu updates

This now works with Flowbite and has simpler implementation. The item method now takes a block so anything can be added e.g. horizontal divider or other control.

We must have coverage through another test, but wouldn't hurt to have a simple unit test for DropdownMenu. This covers the new feature where item now takes an optional block to render any content besides the default link. We can expand this spec further.

* Update styles

* Global Navigation as drawer menu

This is a work in progress. The plan is to have the global navigation always be a vertical drawer menu since the common case is to have many pages. This will allow us to more easily support different viewports as well.

* Disable Style/StringLiterals rule

* Temporarily disable batch actions steps

We already commented out the batch actions feature temporarily as a follow up with migrate over the batch actions form DSL replacement. This is to just appease code coverage for now.