Batch Actions form replacement
The idea is to rely on partials that users declare rather than a DSL so full customization is possible. The batch action now accepts a partial where a user can provide the modal HTML (should be hidden) and link html options so a data attribute can be specified that will open the modal form on click. A delegated form submit event handler is applied for a matching form element by data attribute. The form submission works the same as it did before, providing the same fields.