Update table row click for checkbox toggler
This logic now handles better if a button or anchor is being clicked vs the table row content e.g. text or empty space.

Setting the local type variable we set off this prefer-destructuring warning in ESLint but reviewing the docs it seems unsafe so we'll disable it. We can use it where we want but no need to enforce it everywhere.