Prep Tailwind CSS migration part 4 (#8116)
* Update auto_link to accept html_options

And pass through to link_to method call

* Enable dark mode detection script

We need this for easier testing

* Update page header styles

* Set per-page select to min/auto width

* Add flash styles and icons

* Update header menu icon/button style

* Update dropdown menu button styles

* Disable old global button styles

* Update posts with Tailwind arbitrary value class

* Update scopes styles

* Add anchor default style

* Run yarn build on local rake task run

* Update table row click for checkbox toggler

This logic now handles better if a button or anchor is being clicked vs the table row content e.g. text or empty space.

Setting the local type variable we set off this prefer-destructuring warning in ESLint but reviewing the docs it seems unsafe so we'll disable it. We can use it where we want but no need to enforce it everywhere.

* Set overflow hidden to fix rounded border clipping

The rounded borders around the index table can appear as being clipped off partially so this fixes that so it renders normally.

* Consolidate table thead/th styles

* Remove actions styles (view/edit/delete links)

* Add action item button styles

* Add no-underline to button related components