Update pagination templates and style
We can remove the active_admin_countless theme since all that did differently was not include the first/last links. With out updated pagination theme, we are no longer using first/last links but just next/previous instead. The outermost first and last page number link will be visible so that is its replacement. With that, that means we can remove this extra theme as its no longer needed.

This also updates the pagination locale text to be shorter, matching modern designs.

The paginated_collection component layout has been updated so the first row contains the paging info and links while the second row has the optional per page selector and download links.