Prep Tailwind CSS migration part 5 (#8122)
* Update flowbite and tailwindcss to latest

* Update anchor link style

* Disable old pagination styles

* Update pagination templates and style

We can remove the active_admin_countless theme since all that did differently was not include the first/last links. With out updated pagination theme, we are no longer using first/last links but just next/previous instead. The outermost first and last page number link will be visible so that is its replacement. With that, that means we can remove this extra theme as its no longer needed.

This also updates the pagination locale text to be shorter, matching modern designs.

The paginated_collection component layout has been updated so the first row contains the paging info and links while the second row has the optional per page selector and download links.

* Update comments pagination

* Show comment form before list

Common layout for this is to always have the comment form first and then the list of comments.

* Decrease comment form vertical size to 4 rows

* Use darkest bg color for body in dark mode

* Apply data-table class for index and table_for

Since table_for can be used directly, and its a component vs just a plain table tag, we want to use the same component class for table_for and index_as_table so they are styled the same. We don't want to apply table styles using a tag selector, nor require users to manually add a class if using table_for directly.

* Add arbitrary min widths to columns

This helps testing a responsive table (horizontally scrollable)

* Add dark bg for panel header

* Disable old comments styles

* Update comments styles and layout

* Use singular form for comment delete confirm

* Generate more posts/comments with lorem text

Use some sample lorem sentences to generate random bodies of text for posts and comments. This also bumps the number of each created to help test pagination.

* Replace tailwind gemPath var for local dev

Since we generate an app for local dev in `tmp/development_apps/rails_*` we need to hard code a relative path to the project root. The default value for this is to run `bundle show activeadmin` but that will only work from outside this repo.

* Remove unused range input styles

* Remove unused toggle-bg class styles

This is something Flowbite declared but even before v2 wasn't using in their docs.

* Remove tooltip border styles

We don't want borders on tooltips. The common pattern is its just a solid background color with a shadow. Easier to maintain. Also no need for :after selector on data-popper-arrow as we just need a single psuedo element.

* Adjust padding offset in select from down arrow

* Use "Export:" for sample download text

* Add default viewport meta tag for mobile support

Using latest value per HTML5 Boilerplate

https://github.com/h5bp/html5-boilerplate/blob/main/src/index.html